Add external Ceph cluster support
This patch introduces the API extension to support an external Ceph cluster in
the cinder operator. If Ceph is present (CephMonHosts, CephClusterFSID and
CephClienKey passed as input), the rbd driver is configured and can be enabled
through the `customServiceConfig` volume API key. The common-API structure has
been extended and now includes the CephBackend parameters used to provide all
the information required to setup the "rbd" backend when the config file is
created. The CephBackend belongs to the common-API since it's common to have an
external Ceph cluster serving multiple instances/workloads.
The Ceph Backend also supports:

- CephUser, the rbd user which is reflected within the keyring template and
  cinder.conf

- CephPools as a map of service/CephPoolSpec (definition of the pool)

Those parameters, used as input, are processed to properly build the
Ceph related data structures, mounted later in process to the pods
(mostly cinderVolumes) that are supposed to interact with the external
Ceph cluster.

Signed-off-by: Francesco Pantano <fpantano@redhat.com>