Rework CephBackend and rely to the lib-common
This patch reworks the way cinder-operator defines the CephBackend
and interacts with an external Ceph cluster.
Since most of the OSP related operators are going to share the same
Ceph data structure and functions, we can leave here the logic that
is supposed to configure the service as long as the CephBackend is
defined, importing the required utilities from lib-common.

Signed-off-by: Francesco Pantano <fpantano@redhat.com>