Cleanup httpd.conf
The current WSGI VirtualHost for Cinder API is a big dirty.

We have the same VirtualHost configuration in 2 files,
`10-cinder_wsgi.conf` and `httpd.conf` which makes it hard to know which
one is actually being used.

The reason why it's duplicated is because we currently have the
inclusion of the whole `conf.d` directory where `10-cinder_wsgi.conf`
resides commented to disable SSL.

This patch cleans this up a bit and makes sure that we only have the
VirtualHost configured in one place, `10-cinder_wsgi.conf` and we either
include the configuration files from `conf.d` file or explicitly the
Cinder wsgi configuration.

This removes confusion.