Reduce memory consumption in Cinder services
This patch reduces memory usage on the Cinder Volume and Backup services
by tuning glibc.

The specific tuning consist on disabling the per thread arenas and
disabling dynamic thresholds.

The Cinder Backup service suffers from high water mark memory usage and
uses excessive memory.  As an example just after 10 restore operations
the service uses almost 1GB of RAM and does not ever free it afterwards.
With this patch the memory consumption of the service is reduced down to
almost 130MB and memory consumption during runtime is also halved.

This patch is the equivalent to OpenStack patches:

- DevStack: https://review.opendev.org/c/openstack/devstack/+/845805
- TripleO: https://review.opendev.org/c/openstack/tripleo-common/+/845807

Information on glibc tuning is available at:
https://www.gnu.org/software/libc/manual/html_node/Memory-Allocation-Tunables.html