Run backup container as privileged
We currently run the volume service as privileged because it has to
attach/detach volumes and it also has to access devices under /dev.

The backup services has the same requirements, as it attached/detaches
volumes for backups and restores, so it should have the same privileges
as the cinder-volume container.

This patch updates the cinder backup statefulset with privileged set to
true.