Use keystone catalog info for Glance
Our current cinder.conf template hardcodes the Glance server, which
means that operators will need to set this manually.

This patch changes this to use the Keystone catalog instead, telling
cinder to check for Glance's internal endpoint to get the actual
address.

This seems like a sane default, because even if it is slower due to the
additional request to keystone, it has the benefit of always working and
automatically adapting to any changes.

In the future we could make the operator set the glance_api_servers
itself checking the OpenShift service and then add a watcher to it so it
can react if it were to change.

That would give users the best of both worlds, speed and automatic
updates of cinder.conf on changes.