Makefile: Ensure there's a kubectl
This project is intended for OpenShift and doesn't look to support
Kubernetes, so we should not use kubectl commands in our Makefile.

Unfortunately we use the operator-sdk, which uses kubectl, so we need to
somehow have kubectl.

This patch adds a check to see if there is a kubectl in the path and if
it's not there it tries to create a symlink to oc.

This will work not help if we have the oc command installed in a
directory that requires higher privileges for writing, but it definitely
helps in development with CRC where oc lives in a directory within
~/.crc/