Mount storage related dirs in backup
The backup service is missing some host directories related to accessing
remote storage (for example using iSCSI) and only the volume service is
mounting them.

This is not correct because the backup service also has to attach
volumes using os-brick to back them up and to restore backups.

To fix it this patch moves attachment related volumes/mounts from
`pkg/cindervolume/volumes.go` to `pkg/cinder/volumes.go` and adds a new
parameter its methods `GetVolumes` and `GetVolumeMounts` called
`storageSvc` where caller indicates whether it's a service that uses
storage (volume and backup) or one that doesn't (api and scheduler) to
get the right volumes and mounts.