Simplify resource cleanup
To cleanup the created resources it is ok to just remove the
finalizer which gets set during create. When the owner gets
deleted and the owning resource removes the finalizer. k8s does
does the cleanup of those resources where the owner is missing
and the relevant CR controller calls the cleanup actions, if
any.