Fix the LVM cinder driver
The LVM cinder driver is not working because whenever it calls an LVM
command it hangs [1].

There are 2 ways of fixing it, we can modify the host's
`/etc/lvm/lvm.conf` file and set `udev_sync` and `udev_rules` to `0` or
we can just run the LVM commands on the host using nsenter like we do
with iSCSI and multipath commands.

This patch fixes the LVM cinder driver using the nsenter solution
because it is less intrusive because it doesn't change LVM's becahior on
the host.

The cinder driver calls many different LVM commdands (vgs, lvcreate,
lvdisplay, lvremove, lvextend, etc.), but in the end they are all
symlinks to `/usr/sbin/lvm`, so that's the only command that the
operator will replace in the volume and backup services.

[1]: https://serverfault.com/questions/802766/calling-lvcreate-from-inside-the-container-hangs