Add go.work to deal with multi module dependencies
This patch aligns the cinder-operator to the work already done
in both glance and openstack operator where go.work is added
to deal with multi module dependencies and run linters on API
module.
A go.work default file has been added to the repository, and
the related target is created in the Makefile. However, both
vet and generate targets don't call gowork, and we rely on the
default go.work file.

Signed-off-by: Francesco Pantano <fpantano@redhat.com>