Disable development log mode
This patch disables the zap development logging mode that was the
default in the Cinder operator.

The new code does not set any zap options and only expects to read
options from the command line.

For development purposes it is convenient to run logging not only in
development mode, but also increase the log levels and change the time
encoding from epoc to iso8601 which is easier to read for humans, so 2
new make targets are added, `run-dev` and `docker-build-dev` which are
similar to the `run` and `docker-build` targets.

The `run-dev` target is different from the standard `run` because it
will only build the operator if the binary doesn't exist.