Revamp the service configuration process
The primary goal of this patch is to eliminate using crudini to
merge disparate sets of configuration snippets into a single
cinder.conf file. Instead, we take advantage of oslo's ability to
merge snippets from a config-dir (/etc/cinder/cinder.conf.d), and
leave the original /etc/cinder/cinder.conf file intact.

Configuration snippets are constructed by naming the files in the
numeric order in which we want them merged:
- 00-default.conf will be the contents of the templated cinder.conf
  file provided by the cinder-operator itself.
- 01-secrets.conf will contain secrets extracted from the deployment.
  This includes things such as the rabbit transport and keystone auth
  URLs, database passwords, etc.
- 02-global.conf are global overrides supplied by the user, in the
  customServiceConfig section of the top-level cinder CRD.
- 03-service.conf are any service specific overrides supplied by the
  user in the customServiceConfig section of each service's section
  of the cinder CRD.

As a result of using a config-dir, the startup command for each
service, and its corresponding probe container, is modified to specify
a --config-dir instead of the traditional --config-file.

Lastly, a couple of extraneous files have been deleted, namely the
nfs_shares file, which will be replaced by something more flexible
by using the new extralVol feature.