Support customServiceConfigSecrets
customServiceConfigSecrets provide a mechanism for specifying
service configs via Secrets. Instead of specifying sensitive config
data directly in the customServiceConfig, a cloud admin can place
sensitive data in a Secret, and reference the secret by name in
the service's customServiceConfigSecrets. A list of Secret names
is supported.

Consider a cinder service configured with an NFS backend like this,
where the 'nas_host' and 'nas_share_path' are considered senstive:

      cinderVolumes:
        nfs:
          replicas: 1
          customServiceConfig: |
            [DEFAULT]
            enabled_backends=nfs
            [nfs]
            volume_backend_name=nfs
            volume_driver=cinder.volume.drivers.nfs.NfsDriver
          customServiceConfigSecrets:
          - cinder-volume-nfs-secrets

Along with the following Secret:

  apiVersion: v1
  kind: Secret
  metadata:
    name: cinder-volume-nfs-secrets
  type: Opaque
  stringData:
    nfs-secrets.conf: |
      [nfs]
      nas_host=192.168.130.1
      nas_share_path=/var/nfs/cinder

During the container init process, the contents of each service's
customServiceConfigSecrets (which are optional) are concatenated
into a single file in the directory where all service configs
are located (e.g. /etc/cinder/cinder.conf/cinder.conf.d). In the
Init container, each Secret is mounted in its own directory, which
ensures two secrets that happen to share the same file name
("nfs-secrets.conf" in the example above) do not collide.