Fix operator-lint 0.2.1 findings
This patch fixes the operator-lint target to actually work in
cinder-operator. Pins the operator-lint version to 0.2.1 and fixes the
finding.

The fixed finding is about mixing Required and omitempty in the same
field def. See https://github.com/gibizer/operator-lint/tree/main/linters/crd/C002

Note that this makes the Secret and the DatabaseInstance fields properly
required.