Infer enabled_backends if it's not specified in the CR
The CR for cinder-volume services describe the backends, which are
expected to be listed in the [DEFAULT]/enabled_backends. In order
to make things easier on the cloud admin, the cinder-operator can
now infer the enabled_backends by scanning the customServiceConfig
that defines the backends.

If the customServiceConfig does specify the enabled_backends, then
that value is honored. The enabled_backends is inferred only if
it's not specified in the CR.

Note that this is not equivalent to a transforming or defaulting
webhook, because the CR is not affected. It only affects the final
settings in the generated cinder.conf.