Backend config samples
This patch adds a Ceph backend sample and creates the skeleton for other
backend samples.

The idea is for people to be able to just look at one of the files
in `config/samples/backends/*/backend.yaml` and see just the relevant
bits to configure the Cinder backend.

Other files may be added in there, since credentials will be provided in
a different manifest and then referenced in the `backend.yaml` file
using the `customServiceConfigSecrets`