LVM Sample
This patch adds a simple mechanism to use the iSCSI LVM Cinder backend
driver for testing.

The sample in `config/samples/backends/lvm-iscsi` includes everything
needed to use the Cinder LVM backend with the iSCSI transport protocol
using multipath (even if there is only 1 path in the multipath) in the
CRC environment.

This sample includes:

- Starting the iscsid service in CRC
- Configuring the multipathd service in CRC
- Starting the multipathd service in CRC
- Creating the LVM VG in the CRC VM that will be used by Cinder
- The Cinder backend configuration

The Cinder backend configuration can be applied with:

  $ oc kustomize config/samples/backends/lvm-iscsi | oc apply -f -

The daemon configuration uses `MachineConfig`, which is the recommended
way to make changes to the nodes so they don't become degraded.

These can be applied during installation, upgrades, or at any point in
time, but installation after applying the `MachineConfig` the CRC VM
will be restarted and it will take a while for the node to be fully
ready again.

Using this iSCSI LVM driver on compute nodes will require the creation
of a network that connects the cinder-volume backend with computes.