NVMe-TCP sample
This patch adds another LVM sample to the `config/samples/backends`, in
this case using the NVMe-TCP protocol.

All NVMe-oF protocols require us to load the `nvme-fabrics` kernel
module.

The sample uses the standard OpenShift mechanism to deploy it using
`MachineConfig` and creating a file in `/etc/modules-load.d` named
`nvme_fabrics.conf`.

It also uses a `MachineConfig` to create the LVM backend.