Trim CinderServices CRDs and optimize extraMounts usage
The goal of this patch is to follow up on the Glance work [1] and give
an alternative approach to what's proposed for Glance in PR#184, where
we tried to trim the CRD as much as possible by injecting extraMounts
from the top-level CR and get notified on a change coming from upper
layer. However, the approach described in the Glance PR#184 "breaks" the
top-down pattern where parameters are specified in the Spec and passed
to the deployment, and in a way it "hides" the data from the requests.
The current patch represents a different implementation to solve the
same problem using a different strategy: it trims less than what's
described in Glance PR#184 but keeps the pattern unchanged. In
particular the "Cinder<Service>Spec" is built within the Cinder
controller, but it's not explicitly included in the Cinder. A
"CinderServiceTemplate", instead, is included and used to define the
top-level input parameters provided and set to the sub-level resources.
By doing this we can still keep the ExtraMounts definition in any
"Cinder<Service>Spec", but we don't replicate the same parameter in the
top-level CRD (as the Spec are not explicitly included).

[1] https://github.com/openstack-k8s-operators/glance-operator/pull/191

Signed-off-by: Francesco Pantano <fpantano@redhat.com>