Deprecate init container and use secrets to store configs
Stop using an init container to generate the snippet files that
configure each cinder service. Now, the contents are rendered
entirely by the controllers, and stored in secrets that can be
mounted directly into the /etc/cinder/cinder.conf.d directory.
Content, including scripts, previously stored in a configmap are
now stored in a corresponding secret.

NOTES:
- This patch deprecates the need for init containers, but they
  will be fully removed in a separate patch.
- The switch from using configmaps to secrets affects a lot of
  function and variable names, and comments. This is cosmetic,
  and will be cleaned up in a separate patch.

A total of 4 config snippet files are generated:
- 00-config.conf contains global settings that are common to every
  cinder service, including ones that are derived from deployment
  secrets (e.g. transport URL, database password, etc.)
- 01-config.conf contains the global customServiceConfig settings
  that apply to every cinder service.
- 02-config.conf contains the customServiceConfig settings that
  are specific to each service.
- 03-config.conf contains secrets specified by each service's
  customServiceConfigSecrets.

The 00 and 01 files are generated by the top-level cinder controller,
and each service's 02 and 03 files are generated by that service's
controller. Each service's controller generates a secret that
contains all 4 files, and uses a volume mount to map them directly
in the /etc/cinder/cinder.conf.d directory. Mounting the files
directly in that directory eliminates the need for kolla to copy
the files from a separate "merged" directory.