Fix Backup getting restarted
Due to a bug in the cinder backup service [1] the service may appear to
be down when doing multiple concurrent operations.

When the service appears as down the probe will detect it and will end
up restarting the service, which aborts all the ongoing operations.

While the bug is getting fixed we increase the service_down_time to 3
minutes. This is not ideal because it also affects the cinder volume
service, but it also prevents human operators from seeing the false down
state.

The alternative would be to change the probing configuration just for
the backup service.

[1]: https://bugs.launchpad.net/cinder/+bug/2026877