Docs: Scheduler performance test results
We now recommend using a single scheduler as it avoids some
inconsistency problems in the scheduling calculations and it's good
enough for most of the deployments.

Since this is a deployment change, as we used to deploy 3 replicas, it's
a good idea to document the change, briefly explain the reasons, and
mention the expected performance of the service with a single replica.

This patch adds the top level explanation with the results of the tests
and also a detailed explanation, in a different file to avoid
cluttering, of the tests that were performed to obtain the results.