Customize 00-default.conf for better logging
This patch represents a follow-up of PR#238 where a sidecar container
cinder-api logging purposes has been introduced. This PR represents
a cleaner way to handle the override for cinder-api, and it relies
on the existing "00-config.conf" template to inject the right
"log_file" path. In general, the top-level controller generates a
default ConfigFile, which, in most cases (right now for all the
services except cinder-api), is inherited and set in the related k8s
Secret of a given service. However, the cinder-api controller
implements the logic of properly handle overrides via
templateParameters. This logic is at least shared between the
top-level and the cinder-api controllers, and, in the future, it can
be extended to support multiple overrides on additional templates
through the same mechanism, hence a common.go has been created and
can be used to make more deduplication as long as common code is
detected.

Signed-off-by: Francesco Pantano <fpantano@redhat.com>