Clean up GetVolumes() functions
The unused "secretNames" arg is removed from several GetVolumes()
functions. Its need was eliminated when the config process was
overhauled to not use init containers.

For the cinder-backup and cinder-volume versions of the function,
two volumes and their mounts were moved to the top level
cinder.GetVolumes() and GetVolumeMounts() functions, which already
has specific support for those services.