Support service specific config snippets
Add support for a new service specific config snippet, which is
loaded after the global snippet and before the "custom" global
snippet supplied by the user. This adds the ability for the operator
to override certain global settings on a per-service basis.

The immediate use case is to provide a cleaner way to override the
logging configuration for the cinder-api service.

The use_multipath_for_image_xfer settings in the global snippet have
been moved to new cinder-backup and cinder-volume service specific
snippets.