Backend samples: Use isolated networks
This patch updates the backend samples to use the latest isolated
networks example using Galera for the database.

To avoid having to update our code base every time a change happens in
other operators we now use a reference to the sample from the
openstack-operator repository and add a layer with our overrides.

Right now the only override is disabling Manila and enabling debug log
levels for Cinder services.