Allow cloning when glance uses cinder as backend
When Glance use Cinder as a backend we want create volume from image to
be more efficient and clone the original glance volume where the image
is instead of downloading the Glance image to local disk and then copy
the data to the new volume.

For that we need to tell Cinder that it's OK to do that with the Cinder
volumes using the `allowed_direct_url_schemes` configuration option.