Make keystone check the service tokens
When Cinder receives a service token in the header KeyStone middleware
checks that the token is valid, but it doesn't automatically check that
the role of the token is actually that of a service, for that it needs
to be explicitly told so.

This patch instructs KeyStone middleware to check the role of the
service token when it is present in a request using the
`service_token_roles_required` configuration option.