Backend samples: Simplify steps
This patch changes how we use examples to make it simpler to deploy
them.

It's simple because we are basically following the same steps we would
do for any deployment that uses `install_yamls` with the exception of
generating the YAML and exporting the variable to reference it.

The patch also makes 2 small clarifications in the README.md file