Samples: NVMe-oF kernel module loads
In our tests it looked like it was enough to load the nvme-fabrics
kernel module, as this seemed to load the nvme-tcp kernel module as
needed, but we were told some vendors had to load the other modules for
the OpenStack CI to work as expected.

Just to be on the safe side we add the nvme-tcp and nvme-rdma kernel
modules to the list to load.